+++ api/app/redis/minhash.py
@@ -21,6 +21,14 @@ def schedule_add_genome_signature(sample_id: str, signature) -> SubmittedJob | s
     return SubmittedJob(id=job.id, task=task)
 
 
+def schedule_remove_genome_signature(sample_id: str) -> SubmittedJob | str:
+    """Schedule adding signature to index."""
+    task = "minhash_service.tasks.remove_signature"
+    job = redis.minhash.enqueue(task, sample_id=sample_id, job_timeout="30m")
+    LOG.debug("Submitting job, %s to %s", task, job.worker_name)
+    return SubmittedJob(id=job.id, task=task)
+
+
 def schedule_add_genome_signature_to_index(
     sample_ids: List[str], depends_on: List[str] = None, **enqueue_kwargs
 ) -> SubmittedJob:
@@ -29,7 +37,36 @@ def schedule_add_genome_signature_to_index(
 
     The job can depend on the completion of previous jobs by providing a job_id
     """
-    task = "minhash_service.tasks.index"
+    task = "minhash_service.tasks.add_to_index"
+    submit_kwargs = {
+        "retry": Retry(max=3, interval=60),
+        **enqueue_kwargs,
+    }  # default retry 3 times, 60 in between
+    # make job depend on the job of others
+    if depends_on is not None:
+        submit_kwargs["depends_on"] = Dependency(
+            jobs=depends_on,
+            allow_failure=False,  # allow if dependent job fails
+            enqueue_at_front=True,  # put dependents at front of queue
+        )
+
+    # submit job
+    job = redis.minhash.enqueue(
+        task, sample_ids=sample_ids, job_timeout="30m", **submit_kwargs
+    )
+    LOG.debug("Submitting job, %s to %s", task, job.worker_name)
+    return SubmittedJob(id=job.id, task=task)
+
+
+def schedule_remove_genome_signature_from_index(
+    sample_ids: List[str], depends_on: List[str] = None, **enqueue_kwargs
+) -> SubmittedJob:
+    """
+    Schedule removing signature from index.
+
+    The job can depend on the completion of previous jobs by providing a job_id
+    """
+    task = "minhash_service.tasks.remove_from_index"
     submit_kwargs = {
         "retry": Retry(max=3, interval=60),
         **enqueue_kwargs,